PSR-266 Moved WhoReceivedLoanPage into a common package
