+++ app/controllers/nonsipp/loansmadeoroutstanding/LoansListController.scala
@@ -32,6 +32,7 @@ import models.requests.DataRequest
 import models.{Mode, Money, NormalMode, ReceivedLoanType}
 import navigation.Navigator
 import pages.nonsipp.accountingperiod.AccountingPeriodListPage
+import pages.nonsipp.common.WhoReceivedLoans
 import pages.nonsipp.loansmadeoroutstanding._
 import play.api.data.Form
 import play.api.i18n.MessagesApi
