+++ app/controllers/nonsipp/loansmadeoroutstanding/RecipientSponsoringEmployerConnectedPartyController.scala
@@ -23,6 +23,7 @@ import models.SchemeId.Srn
 import models.requests.DataRequest
 import models.{Mode, ReceivedLoanType, SponsoringOrConnectedParty}
 import navigation.Navigator
+import pages.nonsipp.common.WhoReceivedLoanPage
 import pages.nonsipp.loansmadeoroutstanding._
 import play.api.data.Form
 import play.api.i18n.{I18nSupport, MessagesApi}
