+++ app/controllers/nonsipp/loansmadeoroutstanding/RemoveLoanController.scala
@@ -25,6 +25,7 @@ import models.SchemeId.Srn
 import models.requests.DataRequest
 import models.{Mode, ReceivedLoanType}
 import navigation.Navigator
+import pages.nonsipp.common.WhoReceivedLoanPage
 import pages.nonsipp.loansmadeoroutstanding._
 import play.api.data.Form
 import play.api.i18n.MessagesApi
