+++ test/controllers/nonsipp/loansmadeoroutstanding/RecipientSponsoringEmployerConnectedPartyControllerSpec.scala
@@ -22,6 +22,7 @@ import controllers.nonsipp.loansmadeoroutstanding.RecipientSponsoringEmployerCon
 import eu.timepit.refined.refineMV
 import forms.RadioListFormProvider
 import models.{NormalMode, ReceivedLoanType, RecipientDetails, SponsoringOrConnectedParty, UserAnswers}
+import pages.nonsipp.common.WhoReceivedLoanPage
 import pages.nonsipp.loansmadeoroutstanding._
 import views.html.RadioListView
